app/requirements.txt
asgiref==3.8.1
black==22.1.0
blue==0.9.1
click==8.1.7
colorama==0.4.6
crispy-bootstrap5==2024.2
cssbeautifier==1.15.1
Django==5.0.4
django-crispy-forms==2.1
djlint==1.34.1
EditorConfig==0.12.4
flake8==4.0.1
html-tag-names==0.1.2
html-void-elements==0.1.0
jsbeautifier==1.15.1
json5==0.9.25
mccabe==0.6.1
mypy-extensions==1.0.0
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.0
psycopg[binary]==3.1.18
pycodestyle==2.8.0
pyflakes==2.4.0
python-decouple==3.8
PyYAML==6.0.1
regex==2023.12.25
six==1.16.0
sqlparse==0.5.0
tomli==2.0.1
tqdm==4.66.2
typing_extensions==4.11.0
